setup test finishing. coverage bad
